+++ app/db.js
@@ -9,13 +9,18 @@ exports: module.exports = {
 }
 
 let HEADERS = null
+let SVC_DB_URL = ''
+let ADAPTER_DB_URL = ''
 
 async function initDB( cfg ) {
   HEADERS = { headers: { 
     'app-id'     : cfg.LOWCODE_DB_API_ID,
     'app-secret' : cfg.LOWCODE_DB_API_KEY
   }}
-  
+  let appURL = cfg.LOWCODE_DB_API_URL +'/entity/'+ cfg.LOWCODE_DB_ROOTSCOPE + '/lowcode-integrator/0.1.0'
+  SVC_DB_URL = appURL + '/LCI-Service'
+  ADAPTER_DB_URL = appURL + '/LCI-Adapter'
+
   await prepareDbApp( cfg )
 }
 
@@ -38,6 +43,14 @@ async function prepareDbApp( cfg ) {
       APP.scopeId =  cfg.LOWCODE_DB_ROOTSCOPE
       let result = await axios.post( appURL, APP, HEADERS )
       log.info( 'INIT DB create app', result.status, result.statusText )
+
+      result = await axios.post( 
+        cfg.LOWCODE_DB_API_URL +'/state/'+ cfg.LOWCODE_DB_ROOTSCOPE + '/LCI-Adapter-Status', 
+        ADAPTER_STATE_MODEL, 
+        HEADERS 
+        )
+      log.info( 'INIT DB create adapter state model', result.status, result.statusText )
+
     } 
 
   } catch ( exc ) {
@@ -48,13 +61,10 @@ async function prepareDbApp( cfg ) {
 
 // ============================================================================
 
-async function registerPod( podId, cfg ) {
+async function registerPod( podId, podMode, podURL ) {
   try {
-    let appURL = cfg.LOWCODE_DB_API_URL +'/entity/'+ cfg.LOWCODE_DB_ROOTSCOPE + '/lowcode-integrator/0.1.0'
-    let dtaUrl = appURL + '/LCI-Service'
-    log.info( 'INIT POD', dtaUrl )
-
-    let podList = await axios.get( dtaUrl, HEADERS )
+    log.info( 'INIT POD', SVC_DB_URL )
+    let podList = await axios.get( SVC_DB_URL, HEADERS )
     for ( let uid in podList.data ) {
       if (  podList.data[ uid ].PodId == podId ) {
         log.info( 'INIT POD: already registered', uid )
@@ -64,18 +74,29 @@ async function registerPod( podId, cfg ) {
 
     let podRec = { 
       PodId  : podId,
-      Mode   : cfg.POD_MODE,
-      ApiUrl : cfg.POD_URL,
+      Mode   : podMode,
+      ApiUrl : podURL,
       State  : 'Initialized'
     }
 
-    let result = await axios.post( dtaUrl, podRec, HEADERS )
+    let result = await axios.post( SVC_DB_URL, podRec, HEADERS )
     log.info( 'REGISTER POD', result.data )
+
+    if ( podMode == 'ADAPTER' ) {
+      let registerURL =  ADAPTER_DB_URL+'/state/null/Register'
+      log.info( 'REGISTER ADAPTER', registerURL )
+      let adapterRec = { 
+        AdapterName  : podId,
+        PodId : podId
+      }
+      let result2 = await axios.post( registerURL, adapterRec, HEADERS )
+      log.info( 'REGISTER ADAPTER', result2.data )
+    }
     return result.data.id
 
   } catch ( exc ) {
     log.warn( 'INIT POD', exc.message )
-    process.exit()
+    if ( podMode == 'MANAGER' ) { process.exit() }
   }
 }
 
@@ -129,38 +150,55 @@ const APP = {
           "label": "Adapter Name",
           "type": "String",
           "noEdit": true,
-          "noDelete": true
+          "noDelete": true,
+          "stateTransition": {
+            "null_Register": true
+          }
         },
         "PodId": {
           "label": "Pod Id",
           "type": "String",
           "noEdit": true,
-          "noDelete": true
+          "noDelete": true,
+          "stateTransition": {
+            "null_Register": true
+          }
         },
         "DataInput": {
           "type": "SelectRef",
           "noEdit": true,
           "noDelete": true,
           "selectRef": "2095/lowcode-integrator/0.1.0/LCI-Resource",
-          "label": "Data Input"
+          "label": "Data Input",
+          "stateTransition": {
+            "ConfigPending_Configure": true
+          }
         },
         "Code": {
           "label": "Code",
           "type": "Text",
           "noEdit": true,
           "noDelete": true,
-          "lines": 3
+          "lines": 3,
+          "stateTransition": {
+            "ConfigPending_Configure": true
+          }
         },
         "DataOutput": {
           "label": "Data Output",
           "type": "SelectRef",
           "noEdit": true,
           "noDelete": true,
-          "selectRef": "2095/lowcode-integrator/0.1.0/LCI-Resource"
+          "selectRef": "2095/lowcode-integrator/0.1.0/LCI-Resource",
+          "stateTransition": {
+            "ConfigPending_Configure": true
+          }
         }
       },
       "noDelete": true,
-      "noEdit": true
+      "noEdit": true,
+      "stateModel": "LCI-Adapter-Status",
+      "stateTransition": {}
     },
     "LCI-Resource": {
       "title": "Resource",
@@ -222,5 +260,113 @@ const APP = {
     "appUser"
   ],
   "scope": {
+  }
+}
+
+const ADAPTER_STATE_MODEL =  {
+  "id": "2095/LCI-Adapter-Status",
+  "scopeId": "2095",
+  "state": {
+    "null": {
+      "actions": {
+        "Register": {
+          "to": "ConfigPending",
+          "label": "Register",
+          "apiManaged": true
+        }
+      }
+    },
+    "ConfigPending": {
+      "actions": {
+        "Start": {
+          "to": "Started",
+          "label": "Start",
+          "apiManaged": true
+        },
+        "Configure": {
+          "to": "ConfigPending",
+          "label": "Configure",
+          "apiManaged": true,
+          "line": [
+            {
+              "x": 500,
+              "y": 100
+            },
+            {
+              "x": 500,
+              "y": 150
+            },
+            {
+              "x": 400,
+              "y": 150
+            }
+          ],
+          "labelPos": {
+            "x": 420,
+            "y": 150
+          }
+        }
+      },
+      "label": "Config Pending",
+      "x": 400,
+      "y": 100
+    },
+    "Started": {
+      "actions": {
+        "Stop": {
+          "to": "Stopped",
+          "label": "Stop",
+          "apiManaged": true
+        }
+      },
+      "x": 700,
+      "y": 100,
+      "label": "Started"
+    },
+    "Stopped": {
+      "actions": {
+        "Restart": {
+          "to": "Started",
+          "label": "Restart",
+          "line": [
+            {
+              "x": 1000,
+              "y": 150
+            },
+            {
+              "x": 700,
+              "y": 150
+            }
+          ],
+          "apiManaged": true,
+          "labelPos": {
+            "x": 870,
+            "y": 150
+          }
+        },
+        "Reconfigure": {
+          "to": "ConfigPending",
+          "label": "Reconfigure",
+          "line": [
+            {
+              "x": 1000,
+              "y": 200
+            },
+            {
+              "x": 400,
+              "y": 200
+            }
+          ],
+          "labelPos": {
+            "x": 870,
+            "y": 200
+          },
+          "apiManaged": true
+        }
+      },
+      "x": 1000,
+      "y": 100,
+      "label": "Stopped"
+    }
   }
 }