+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.0.2
 - init app
 - register pod
+- adapter state model + init adapter registration